feat: Introduce Service Firewall
Part of #5
