--- config/foo.yml
@@ -1,1 +1,0 @@
-name: 'hello'
